requirements.txt
pandas
requests
ccxt
python-dotenv
